requirements.txt
python-dotenv
django
langchain-openai
langchain
openai
google-generativeai
google-auth
google-auth-oauthlib
google-api-python-client
livekit-agents
livekit-plugins
livekit
deepgram-sdk
elevenlabs
pytz
tenacity
python-dateutil
structlog
asyncio
gunicorn
